requirements.txt
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
opentelemetry-semantic-conventions>=0.41b0
pyyaml>=6.0
pytest>=8.0
